+++ sdk/dynatrace/Azure.ResourceManager.Dynatrace/tests/Scenario/MonitorResourceOperationsTests.cs
@@ -8,7 +8,7 @@ using NUnit.Framework;
 
 namespace Azure.ResourceManager.Dynatrace.Tests
 {
-    public class MonitorResourceOperationsTests : MonitorTestBase
+    public class MonitorResourceOperationsTests : DynatraceManagementTestBase
     {
         public MonitorResourceOperationsTests(bool isAsync)
             : base(isAsync)//, RecordedTestMode.Record)
@@ -17,17 +17,12 @@ namespace Azure.ResourceManager.Dynatrace.Tests
 
         [TestCase]
         [RecordedTest]
+        [Ignore("Creation of dynatrace monitor will fail and need fix.")]
         public async Task Get()
         {
-            var collection = await GetMonitorResourceCollectionAsync();
             string monitorName = Recording.GenerateAssetName("testDT-");
-
-            var input = GetMonitorInput();
-
-            var lro = await collection.CreateOrUpdateAsync(WaitUntil.Completed, monitorName, input);
-            DynatraceMonitorResource monitorResource1 = lro.Value;
-
-            DynatraceMonitorResource monitorResource2 = await collection.GetAsync(monitorName);
+            DynatraceMonitorResource monitorResource1 = await CreateMonitorResourceAsync(monitorName);
+            DynatraceMonitorResource monitorResource2 = await monitorResource1.GetAsync();
 
             Assert.AreEqual(monitorResource1.Data.Name, monitorResource2.Data.Name);
             Assert.AreEqual(monitorResource1.Data.Id, monitorResource2.Data.Id);
@@ -38,6 +33,7 @@ namespace Azure.ResourceManager.Dynatrace.Tests
 
         [TestCase]
         [RecordedTest]
+        [Ignore("Creation of dynatrace monitor will fail and need fix.")]
         public async Task GetVMPayload()
         {
             string monitorName = Recording.GenerateAssetName("testDT-");
